// OxpeckerOpenApi.BugTest/generated-client/models/animal-color.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Oxpecker OpenAPI Bug Test API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * F# union type with values: Brown, Black, White, Spotted
 */

export enum AnimalColor {
    Brown = 'Brown',
    Black = 'Black',
    White = 'White',
    Spotted = 'Spotted'
}