// OxpeckerOpenApi.BugTest/generated-client/models/shape.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Oxpecker OpenAPI Bug Test API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ShapeOneOf } from './shape-one-of';
// May contain unused imports in some cases
// @ts-ignore
import type { ShapeOneOf1 } from './shape-one-of1';
// May contain unused imports in some cases
// @ts-ignore
import type { ShapeOneOf2 } from './shape-one-of2';

/**
 * @type Shape
 * F# discriminated union type with cases: Circle, Rectangle, Triangle
 */
export type Shape = { type: 'Circle' } & ShapeCircle | { type: 'Rectangle' } & ShapeRectangle | { type: 'Triangle' } & ShapeTriangle;